Got secret copying working over ssh.

+++ Threax.Provision.CheapAzure/Services/SshCredsManager.cs
@@ -16,35 +16,40 @@ namespace Threax.Provision.CheapAzure.Services
         private readonly IKeyVaultManager keyVaultManager;
         private readonly ICredentialLookup credentialLookup;
         private readonly IProcessRunner processRunner;
+        private readonly IKeyVaultAccessManager keyVaultAccessManager;
         private String publicKeyFile;
         private String privateKeyFile;
         private String vmUser;
 
-        public SshCredsManager(Config config, IKeyVaultManager keyVaultManager, ICredentialLookup credentialLookup, IProcessRunner processRunner)
+        public SshCredsManager(Config config, IKeyVaultManager keyVaultManager, ICredentialLookup credentialLookup, IProcessRunner processRunner, IKeyVaultAccessManager keyVaultAccessManager)
         {
             this.config = config;
             this.keyVaultManager = keyVaultManager;
             this.credentialLookup = credentialLookup;
             this.processRunner = processRunner;
+            this.keyVaultAccessManager = keyVaultAccessManager;
         }
 
         public void Dispose()
         {
-            try
+            if (!string.IsNullOrEmpty(publicKeyFile))
             {
-                File.Delete(publicKeyFile);
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"An {ex.GetType().Name} occured cleaning up public key file '{publicKeyFile}'. Message: {ex.Message}\n{ex.StackTrace}");
-            }
-            try
-            {
-                File.Delete(privateKeyFile);
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"An {ex.GetType().Name} occured cleaning up private key file '{privateKeyFile}'. Message: {ex.Message}\n{ex.StackTrace}");
+                try
+                {
+                    File.Delete(publicKeyFile);
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine($"An {ex.GetType().Name} occured cleaning up public key file '{publicKeyFile}'. Message: {ex.Message}\n{ex.StackTrace}");
+                }
+                try
+                {
+                    File.Delete(privateKeyFile);
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine($"An {ex.GetType().Name} occured cleaning up private key file '{privateKeyFile}'. Message: {ex.Message}\n{ex.StackTrace}");
+                }
             }
         }
 
@@ -60,7 +65,7 @@ namespace Threax.Provision.CheapAzure.Services
             return publicKey;
         }
 
-        public async Task RunSshCommand(String command, int safeExitCode = 0)
+        public async Task<int> RunSshCommand(String command)
         {
             if (String.IsNullOrEmpty(config.VmSshHost))
             {
@@ -69,11 +74,7 @@ namespace Threax.Provision.CheapAzure.Services
 
             var privateKeyPath = await LoadKeysAndGetSshPrivateKeyPath();
             var startInfo = new ProcessStartInfo("ssh", $"-i \"{privateKeyPath}\" -t \"{vmUser}@{config.VmSshHost}\" \"{command}\"");
-            var exitCode = processRunner.RunProcessWithOutput(startInfo);
-            if (exitCode != safeExitCode)
-            {
-                throw new InvalidOperationException($"Error running command '{command}'. Exit code was '{exitCode}' expected '{safeExitCode}'");
-            }
+            return processRunner.RunProcessWithOutput(startInfo);
         }
 
         public async Task CopySshFile(String file, String dest)
@@ -92,27 +93,9 @@ namespace Threax.Provision.CheapAzure.Services
             }
         }
 
-        public async Task CopySshStringToFile(String content, String dest)
-        {
-            var tempFile = Path.GetTempFileName();
-            try
-            {
-                File.WriteAllText(tempFile, content);
-                await CopySshFile(tempFile, dest);
-            }
-            finally
-            {
-                //Any exceptions here are intentionally left to bubble up
-                if (File.Exists(tempFile))
-                {
-                    File.Delete(tempFile);
-                }
-            }
-        }
-
         public String PublicKeySecretName => $"{config.VmAdminBaseKey}-ssh-public-key";
 
-        public String PrivateKeySecretName => $"{config.VmAdminBaseKey}-ssh-public-key";
+        public String PrivateKeySecretName => $"{config.VmAdminBaseKey}-ssh-private-key";
 
         private String GetUserHomePath()
         {
@@ -128,10 +111,22 @@ namespace Threax.Provision.CheapAzure.Services
             if (privateKeyFile == null)
             {
                 publicKeyFile = Path.Combine(GetUserHomePath(), ".threaxprovision", "azure-ssh.pub");
+                if (!Directory.Exists(Path.GetDirectoryName(publicKeyFile)))
+                {
+                    Directory.CreateDirectory(Path.GetDirectoryName(publicKeyFile));
+                }
+                
                 privateKeyFile = Path.Combine(GetUserHomePath(), ".threaxprovision", "azure-ssh");
+                if (!Directory.Exists(Path.GetDirectoryName(privateKeyFile)))
+                {
+                    Directory.CreateDirectory(Path.GetDirectoryName(privateKeyFile));
+                }
+
+                await keyVaultAccessManager.Unlock(config.InfraKeyVaultName, config.UserId);
 
                 var publicKey = await keyVaultManager.GetSecret(config.InfraKeyVaultName, PublicKeySecretName);
                 var privateKey = await keyVaultManager.GetSecret(config.InfraKeyVaultName, PrivateKeySecretName);
+                privateKey = privateKey.Replace("**lf**", "\n"); //Replace all placeholder line feeds with real line feeds
 
                 File.WriteAllText(publicKeyFile, publicKey);
                 File.WriteAllText(privateKeyFile, privateKey);
